requirements_frozen.txt
# generate me by (&manual cleaning): pip freeze > requirements_frozen.txt
fair-esm==2.0.0
h5py==3.10.0
hdbscan==0.8.33
matplotlib==3.8.4
pandas==1.5.3
qdrant-client==1.7.1
scikit-learn==1.4.0
taxoniq==1.0.1
tqdm==4.66.1
requests==2.31.0
plotly==5.18.0
biopython==1.84
dash-bootstrap-components==1.6.0